feat(android): Download/Auralis MediaStore dual-save + filesystem verification (v2.5.11)
- android_downloads.rs: publish_to_downloads via MediaStore.Downloads IS_PENDING (API29+) with legacy Environment+MediaScanner fallback, cached_copy_for_path for player fallback
- downloader.rs: after complete publish copy to Download/Auralis (dual-save, non-fatal)
- player.rs: play() fallback to cached_copy_for_path when File::open fails on Download/Auralis/content://
- library.rs: keep scan sandboxed to avoid duplicate entries
- settings.rs: DownloadSettings.use_system_downloads (default true) serde default_true
- settings.html/views.js: toggle Save to system Downloads (Download/Auralis)
- e2e_download_test.js/run_emulator_test.sh/verify_downloads_mediastore.sh: ls /storage/emulated/0/Download/Auralis + content query is_pending 0 + internal dual-save check
- bump 2.5.10 -> 2.5.11 Cargo.toml/Cargo.lock/tauri.conf.json/package.json + AGENTS §4.5/PROJECT §9/README/upgrade_test docs
- fixes: android_downloads clippy doc/bind/unnecessary_cast + fmt

+++ src/lib.rs
@@ -438,5 +438,5 @@ pub extern "system" fn JNI_OnLoad(vm: jni::JavaVM, _reserved: *mut std::ffi::c_v
     android_jni::INITIAL_VM.store(vm.get_java_vm_pointer() as *mut c_void, Ordering::SeqCst);
     android_jni::try_seed(&vm);
 
-    jni::sys::JNI_VERSION_1_6 as i32
+    jni::sys::JNI_VERSION_1_6
 }
